ci: Adjust AppVeyor to use parallel building script
Adjusts the build script for AppVeyor to be windows only, and also to use my parallel build runner from obs-stream-effects, which effectively reduces the build time by up to 10% as less time is spent waiting and all resources are being used all the time. There are now three stages of building, configure -> build -> package. Artifacts are also uploaded and can be used for direct linking if desired.