UI: Split main function from OBSApp class
To increase maintainability of the code, many smaller files are much better than one huge file. Only code directly related to the task should be in the same file, all other code should be offloaded into additional files that are included.

This also decreases compile time as many of the generated object files never change, instead of requiring a load of the actual file to see if it changed. Downside is that on file systems with a large block size, these files will take more space.